+++ packages/abi/lib.esm/fragments.js
@@ -27,7 +27,7 @@ function checkModifier(type, name) {
 function parseParamType(param, allowIndexed) {
     let originalParam = param;
     function throwError(i) {
-        throw new Error("unexpected character '" + originalParam[i] + "' at position " + i + " in '" + originalParam + "'");
+        logger.throwArgumentError(`unexpected character at position ${i}`, "param", param);
     }
     param = param.replace(/\s/g, " ");
     function newNode(parent) {
@@ -166,7 +166,7 @@ function parseParamType(param, allowIndexed) {
         }
     }
     if (node.parent) {
-        throw new Error("unexpected eof");
+        logger.throwArgumentError("unexpected eof", "param", param);
     }
     delete parent.state;
     if (node.name === "indexed") {
@@ -204,7 +204,9 @@ const paramTypeArray = new RegExp(/^(.*)\[([0-9]*)\]$/);
 export class ParamType {
     constructor(constructorGuard, params) {
         if (constructorGuard !== _constructorGuard) {
-            throw new Error("use fromString");
+            logger.throwError("use fromString", Logger.errors.UNSUPPORTED_OPERATION, {
+                operation: "new ParamType()"
+            });
         }
         populate(this, params);
         let match = this.type.match(paramTypeArray);
@@ -318,7 +320,9 @@ function parseParams(value, allowIndex) {
 export class Fragment {
     constructor(constructorGuard, params) {
         if (constructorGuard !== _constructorGuard) {
-            throw new Error("use a static from method");
+            logger.throwError("use a static from method", Logger.errors.UNSUPPORTED_OPERATION, {
+                operation: "new Fragment()"
+            });
         }
         populate(this, params);
         this._isFragment = true;
@@ -365,7 +369,7 @@ export class Fragment {
         else if (value.split("(")[0].trim() === "constructor") {
             return ConstructorFragment.fromString(value.trim());
         }
-        throw new Error("unknown fragment");
+        return logger.throwArgumentError("unsupported fragment", "value", value);
     }
     static isFragment(value) {
         return !!(value && value._isFragment);
@@ -410,7 +414,7 @@ export class EventFragment extends Fragment {
             return value;
         }
         if (value.type !== "event") {
-            throw new Error("invalid event object - " + value.type);
+            logger.throwArgumentError("invalid event object", "value", value);
         }
         return new EventFragment(_constructorGuard, {
             name: verifyIdentifier(value.name),
@@ -422,7 +426,7 @@ export class EventFragment extends Fragment {
     static fromString(value) {
         let match = value.match(regexParen);
         if (!match) {
-            throw new Error("invalid event: " + value);
+            logger.throwArgumentError("invalid event string", "value", value);
         }
         let anonymous = false;
         match[3].split(" ").forEach((modifier) => {
@@ -452,10 +456,10 @@ function parseGas(value, params) {
     let comps = value.split("@");
     if (comps.length !== 1) {
         if (comps.length > 2) {
-            throw new Error("invalid signature");
+            logger.throwArgumentError("invalid human-readable ABI signature", "value", value);
         }
         if (!comps[1].match(/^[0-9]+$/)) {
-            throw new Error("invalid signature gas");
+            logger.throwArgumentError("invalid human-readable aBI signature gas", "value", value);
         }
         params.gas = BigNumber.from(comps[1]);
         return comps[0];
@@ -504,14 +508,14 @@ function verifyState(value) {
         result.constant = (result.stateMutability === "view" || result.stateMutability === "pure");
         if (value.constant != null) {
             if ((!!value.constant) !== result.constant) {
-                throw new Error("cannot have constant function with mutability " + result.stateMutability);
+                logger.throwArgumentError("cannot have constant function with mutability " + result.stateMutability, "value", value);
             }
         }
         // Set (and check things are consistent) the payable property
         result.payable = (result.stateMutability === "payable");
         if (value.payable != null) {
             if ((!!value.payable) !== result.payable) {
-                throw new Error("cannot have payable function with mutability " + result.stateMutability);
+                logger.throwArgumentError("cannot have payable function with mutability " + result.stateMutability, "value", value);
             }
         }
     }
@@ -519,7 +523,7 @@ function verifyState(value) {
         result.payable = !!value.payable;
         // If payable we can assume non-constant; otherwise we can't assume
         if (value.constant == null && !result.payable && value.type !== "constructor") {
-            throw new Error("unable to determine stateMutability");
+            logger.throwArgumentError("unable to determine stateMutability", "value", value);
         }
         result.constant = !!value.constant;
         if (result.constant) {
@@ -529,7 +533,7 @@ function verifyState(value) {
             result.stateMutability = (result.payable ? "payable" : "nonpayable");
         }
         if (result.payable && result.constant) {
-            throw new Error("cannot have constant payable function");
+            logger.throwArgumentError("cannot have constant payable function", "value", value);
         }
     }
     else if (value.constant != null) {
@@ -538,7 +542,7 @@ function verifyState(value) {
         result.stateMutability = (result.constant ? "view" : "payable");
     }
     else if (value.type !== "constructor") {
-        throw new Error("unable to determine stateMutability");
+        logger.throwArgumentError("unable to determine stateMutability", "value", value);
     }
     return result;
 }
@@ -581,11 +585,11 @@ export class ConstructorFragment extends Fragment {
             return value;
         }
         if (value.type !== "constructor") {
-            throw new Error("invalid constructor object - " + value.type);
+            logger.throwArgumentError("invalid constructor object", "value", value);
         }
         let state = verifyState(value);
         if (state.constant) {
-            throw new Error("constructor cannot be constant");
+            logger.throwArgumentError("constructor cannot be constant", "value", value);
         }
         return new ConstructorFragment(_constructorGuard, {
             name: null,
@@ -599,11 +603,8 @@ export class ConstructorFragment extends Fragment {
         let params = { type: "constructor" };
         value = parseGas(value, params);
         let parens = value.match(regexParen);
-        if (!parens) {
-            throw new Error("invalid constructor: " + value);
-        }
-        if (parens[1].trim() !== "constructor") {
-            throw new Error("invalid constructor");
+        if (!parens || parens[1].trim() !== "constructor") {
+            logger.throwArgumentError("invalid constructor string", "value", value);
         }
         params.inputs = parseParams(parens[2].trim(), false);
         parseModifiers(parens[3].trim(), params);
@@ -667,7 +668,7 @@ export class FunctionFragment extends ConstructorFragment {
             return value;
         }
         if (value.type !== "function") {
-            throw new Error("invalid function object - " + value.type);
+            logger.throwArgumentError("invalid function object", "value", value);
         }
         let state = verifyState(value);
         return new FunctionFragment(_constructorGuard, {
@@ -686,15 +687,15 @@ export class FunctionFragment extends ConstructorFragment {
         value = parseGas(value, params);
         let comps = value.split(" returns ");
         if (comps.length > 2) {
-            throw new Error("invalid function");
+            logger.throwArgumentError("invalid function string", "value", value);
         }
         let parens = comps[0].match(regexParen);
         if (!parens) {
-            throw new Error("invalid signature");
+            logger.throwArgumentError("invalid function signature", "value", value);
         }
         params.name = parens[1].trim();
-        if (!params.name.match(regexIdentifier)) {
-            throw new Error("invalid identifier: '" + params.name + "'");
+        if (params.name) {
+            verifyIdentifier(params.name);
         }
         params.inputs = parseParams(parens[2], false);
         parseModifiers(parens[3].trim(), params);
@@ -702,7 +703,7 @@ export class FunctionFragment extends ConstructorFragment {
         if (comps.length > 1) {
             let returns = comps[1].match(regexParen);
             if (returns[1].trim() != "" || returns[3].trim() != "") {
-                throw new Error("unexpected tokens");
+                logger.throwArgumentError("unexpected tokens", "value", value);
             }
             params.outputs = parseParams(returns[2], false);
         }
@@ -733,7 +734,7 @@ function verifyType(type) {
 const regexIdentifier = new RegExp("^[A-Za-z_][A-Za-z0-9_]*$");
 function verifyIdentifier(value) {
     if (!value || !value.match(regexIdentifier)) {
-        throw new Error("invalid identifier: '" + value + "'");
+        logger.throwArgumentError(`invalid identifier "${value}"`, "value", value);
     }
     return value;
 }
@@ -757,7 +758,7 @@ function splitNesting(value) {
             else if (c === ")") {
                 depth--;
                 if (depth === -1) {
-                    throw new Error("unbalanced parenthsis");
+                    logger.throwArgumentError("unbalanced parenthsis", "value", value);
                 }
             }
         }