+++ packages/abi/lib/fragments.js
@@ -41,7 +41,7 @@ function checkModifier(type, name) {
 function parseParamType(param, allowIndexed) {
     var originalParam = param;
     function throwError(i) {
-        throw new Error("unexpected character '" + originalParam[i] + "' at position " + i + " in '" + originalParam + "'");
+        logger.throwArgumentError("unexpected character at position " + i, "param", param);
     }
     param = param.replace(/\s/g, " ");
     function newNode(parent) {
@@ -180,7 +180,7 @@ function parseParamType(param, allowIndexed) {
         }
     }
     if (node.parent) {
-        throw new Error("unexpected eof");
+        logger.throwArgumentError("unexpected eof", "param", param);
     }
     delete parent.state;
     if (node.name === "indexed") {
@@ -218,7 +218,9 @@ var paramTypeArray = new RegExp(/^(.*)\[([0-9]*)\]$/);
 var ParamType = /** @class */ (function () {
     function ParamType(constructorGuard, params) {
         if (constructorGuard !== _constructorGuard) {
-            throw new Error("use fromString");
+            logger.throwError("use fromString", logger_1.Logger.errors.UNSUPPORTED_OPERATION, {
+                operation: "new ParamType()"
+            });
         }
         populate(this, params);
         var match = this.type.match(paramTypeArray);
@@ -334,7 +336,9 @@ function parseParams(value, allowIndex) {
 var Fragment = /** @class */ (function () {
     function Fragment(constructorGuard, params) {
         if (constructorGuard !== _constructorGuard) {
-            throw new Error("use a static from method");
+            logger.throwError("use a static from method", logger_1.Logger.errors.UNSUPPORTED_OPERATION, {
+                operation: "new Fragment()"
+            });
         }
         populate(this, params);
         this._isFragment = true;
@@ -381,7 +385,7 @@ var Fragment = /** @class */ (function () {
         else if (value.split("(")[0].trim() === "constructor") {
             return ConstructorFragment.fromString(value.trim());
         }
-        throw new Error("unknown fragment");
+        return logger.throwArgumentError("unsupported fragment", "value", value);
     };
     Fragment.isFragment = function (value) {
         return !!(value && value._isFragment);
@@ -432,7 +436,7 @@ var EventFragment = /** @class */ (function (_super) {
             return value;
         }
         if (value.type !== "event") {
-            throw new Error("invalid event object - " + value.type);
+            logger.throwArgumentError("invalid event object", "value", value);
         }
         return new EventFragment(_constructorGuard, {
             name: verifyIdentifier(value.name),
@@ -444,7 +448,7 @@ var EventFragment = /** @class */ (function (_super) {
     EventFragment.fromString = function (value) {
         var match = value.match(regexParen);
         if (!match) {
-            throw new Error("invalid event: " + value);
+            logger.throwArgumentError("invalid event string", "value", value);
         }
         var anonymous = false;
         match[3].split(" ").forEach(function (modifier) {
@@ -476,10 +480,10 @@ function parseGas(value, params) {
     var comps = value.split("@");
     if (comps.length !== 1) {
         if (comps.length > 2) {
-            throw new Error("invalid signature");
+            logger.throwArgumentError("invalid human-readable ABI signature", "value", value);
         }
         if (!comps[1].match(/^[0-9]+$/)) {
-            throw new Error("invalid signature gas");
+            logger.throwArgumentError("invalid human-readable aBI signature gas", "value", value);
         }
         params.gas = bignumber_1.BigNumber.from(comps[1]);
         return comps[0];
@@ -528,14 +532,14 @@ function verifyState(value) {
         result.constant = (result.stateMutability === "view" || result.stateMutability === "pure");
         if (value.constant != null) {
             if ((!!value.constant) !== result.constant) {
-                throw new Error("cannot have constant function with mutability " + result.stateMutability);
+                logger.throwArgumentError("cannot have constant function with mutability " + result.stateMutability, "value", value);
             }
         }
         // Set (and check things are consistent) the payable property
         result.payable = (result.stateMutability === "payable");
         if (value.payable != null) {
             if ((!!value.payable) !== result.payable) {
-                throw new Error("cannot have payable function with mutability " + result.stateMutability);
+                logger.throwArgumentError("cannot have payable function with mutability " + result.stateMutability, "value", value);
             }
         }
     }
@@ -543,7 +547,7 @@ function verifyState(value) {
         result.payable = !!value.payable;
         // If payable we can assume non-constant; otherwise we can't assume
         if (value.constant == null && !result.payable && value.type !== "constructor") {
-            throw new Error("unable to determine stateMutability");
+            logger.throwArgumentError("unable to determine stateMutability", "value", value);
         }
         result.constant = !!value.constant;
         if (result.constant) {
@@ -553,7 +557,7 @@ function verifyState(value) {
             result.stateMutability = (result.payable ? "payable" : "nonpayable");
         }
         if (result.payable && result.constant) {
-            throw new Error("cannot have constant payable function");
+            logger.throwArgumentError("cannot have constant payable function", "value", value);
         }
     }
     else if (value.constant != null) {
@@ -562,7 +566,7 @@ function verifyState(value) {
         result.stateMutability = (result.constant ? "view" : "payable");
     }
     else if (value.type !== "constructor") {
-        throw new Error("unable to determine stateMutability");
+        logger.throwArgumentError("unable to determine stateMutability", "value", value);
     }
     return result;
 }
@@ -609,11 +613,11 @@ var ConstructorFragment = /** @class */ (function (_super) {
             return value;
         }
         if (value.type !== "constructor") {
-            throw new Error("invalid constructor object - " + value.type);
+            logger.throwArgumentError("invalid constructor object", "value", value);
         }
         var state = verifyState(value);
         if (state.constant) {
-            throw new Error("constructor cannot be constant");
+            logger.throwArgumentError("constructor cannot be constant", "value", value);
         }
         return new ConstructorFragment(_constructorGuard, {
             name: null,
@@ -627,11 +631,8 @@ var ConstructorFragment = /** @class */ (function (_super) {
         var params = { type: "constructor" };
         value = parseGas(value, params);
         var parens = value.match(regexParen);
-        if (!parens) {
-            throw new Error("invalid constructor: " + value);
-        }
-        if (parens[1].trim() !== "constructor") {
-            throw new Error("invalid constructor");
+        if (!parens || parens[1].trim() !== "constructor") {
+            logger.throwArgumentError("invalid constructor string", "value", value);
         }
         params.inputs = parseParams(parens[2].trim(), false);
         parseModifiers(parens[3].trim(), params);
@@ -701,7 +702,7 @@ var FunctionFragment = /** @class */ (function (_super) {
             return value;
         }
         if (value.type !== "function") {
-            throw new Error("invalid function object - " + value.type);
+            logger.throwArgumentError("invalid function object", "value", value);
         }
         var state = verifyState(value);
         return new FunctionFragment(_constructorGuard, {
@@ -720,15 +721,15 @@ var FunctionFragment = /** @class */ (function (_super) {
         value = parseGas(value, params);
         var comps = value.split(" returns ");
         if (comps.length > 2) {
-            throw new Error("invalid function");
+            logger.throwArgumentError("invalid function string", "value", value);
         }
         var parens = comps[0].match(regexParen);
         if (!parens) {
-            throw new Error("invalid signature");
+            logger.throwArgumentError("invalid function signature", "value", value);
         }
         params.name = parens[1].trim();
-        if (!params.name.match(regexIdentifier)) {
-            throw new Error("invalid identifier: '" + params.name + "'");
+        if (params.name) {
+            verifyIdentifier(params.name);
         }
         params.inputs = parseParams(parens[2], false);
         parseModifiers(parens[3].trim(), params);
@@ -736,7 +737,7 @@ var FunctionFragment = /** @class */ (function (_super) {
         if (comps.length > 1) {
             var returns = comps[1].match(regexParen);
             if (returns[1].trim() != "" || returns[3].trim() != "") {
-                throw new Error("unexpected tokens");
+                logger.throwArgumentError("unexpected tokens", "value", value);
             }
             params.outputs = parseParams(returns[2], false);
         }
@@ -769,7 +770,7 @@ function verifyType(type) {
 var regexIdentifier = new RegExp("^[A-Za-z_][A-Za-z0-9_]*$");
 function verifyIdentifier(value) {
     if (!value || !value.match(regexIdentifier)) {
-        throw new Error("invalid identifier: '" + value + "'");
+        logger.throwArgumentError("invalid identifier \"" + value + "\"", "value", value);
     }
     return value;
 }
@@ -793,7 +794,7 @@ function splitNesting(value) {
             else if (c === ")") {
                 depth--;
                 if (depth === -1) {
-                    throw new Error("unbalanced parenthsis");
+                    logger.throwArgumentError("unbalanced parenthsis", "value", value);
                 }
             }
         }