+++ packages/bignumber/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "bignumber/5.0.0-beta.136";
+export declare const version = "bignumber/5.0.0-beta.137";